[PATCH] aacraid 2.5 update
This contains three changes.  The first removes some #defines that don't
seem to be needed anymore.  The second fixes a problem when CONFIG_LBD
is set and sector_t is u64.  The third fixes some compile warnings
setting a char * in the scsi_pointer struct to a dma_addr_t.  I changed
the usage from the ptr element to the dma_handle element.

It compiles without warnings and I have run some tests on aacraid devices
on the osdl lab machines with the CONFIG_LBD option on and off.