[PATCH] ehci-hcd (2/2): rest of tasklet remove
This is the rest of the work to remove the tasklet: the non-syntax
portions which affect work scheduling.  It's not quite davem's version;
it's got locking updates, which among other things prevent a hang when
the timer kicks in.

This scheduling change is split out from the other parts in case more
problems like that unlink race (fixed in my previous patch) show up.
It doesn't fix (or help fix) any ehci bugs, but simpler code is fine.