kbuild: fix bug in scripts/kallsyms.c
kallsyms.c generates the symbol table in a .S file using the assembler .string
macro.  Unfortunately, the .string macro is implemented in a platform specific
way (it may or may not zero terminate the string).  On parisc, it doesn't zero
terminate, so the symbol table search doesn't work.

The solution is to replace .string with .asciz which is guaranteed to do the
correct thing on all platforms.