[ARM] __put_user, __get_user cleanups and fixups
This fixes the ARM __put_user implementation.  The gcc folk have
agreed that the %Qn and %Rn register notation in asm() statements
should become official, and have been in all compilers which can
be used to build the ARM kernel.

In addition, we remove some historical gunk from when put_user()
and get_user() was inlined on ARM.