ISDN/HiSax: Fix some of the new PnP stuff
I appreciate the adaption of the ISDN drivers to the new PnP layer
(though I don't know why it was actually necessary to break the old
ISAPnP so late in the cycle), but the gratuitious changes to the coding
style were not exactly necessary. So revert things to the
"if (retval) goto err_unwind;" style and fix a couple of other compile
time errors.